% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/immutequality.R
\name{=}
\alias{=}
\title{Immutequality Operator}
\arguments{
\item{x}{a symbol}

\item{value}{the associated value}
}
\description{
Assign \code{value} to the symbol \code{x}, which can then subsequently
be treated equal to \code{value}. This "immutability" is emulated by
assigning an active binding to the symbol \code{x}. If the value is
accessed, it is simply returned by the active binding, but if a value
change is attempted, an error is raised that the symbol cannot be reused.
The immutequality operator can thus be used to promise that the value
associated with a symbol will not change for the remainder of its scope.
}
\examples{
x = 10

print(x)
tryCatch(x <- x*2, error = function(e) message(e))
}

