% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/immutequality.R
\docType{package}
\name{immutequality}
\alias{immutequality}
\alias{immutequality-package}
\title{Use Equality Assignment for Immutability}
\description{
This package overloads the \code{`=`} operator with the
  immutequality operator. It associates a value with a symbol, which can then
  subsequently be treated equal to the value for the remainder of the scope.
  This "immutability" is emulated by assigning an active binding to the
  symbol. If the value is accessed, it is simply returned by the active
  binding, but if a value change is attempted, an error is raised that the
  symbol cannot be reused. The immutequality operator can thus be used to
  promise that the value associated with a symbol will not change for the
  remainder of its scope.
}
\details{
Immutequality
}
\author{
Stefan Milton Bache <stefan@stefanbache.dk>
}

